+++ src/app/intro/routes.js
@@ -1,0 +1,7 @@
+export default [
+  {
+    path: '/intro',
+    name: 'Intro',
+    component: () => import('./Intro' /* webpackChunkName: "intro" */)
+  }
+]