+++ src/app/plant/components/PlantFooter.vue
@@ -1,28 +1,30 @@
 <template>
   <section :class="{ 'plant-footer': true, 'no-modules': noModules }">
-    <button
-      type="button"
-      class="footer-modules plain"
-      @click.prevent="emitShowModules">
-      <feather-box />
-      <span>Manage modules</span>
-    </button>
+    <v-button
+      color="grey"
+      class="footer-modules"
+      @click.native.prevent="emitShowModules">
+      <feather-box slot="icon" />
+      Manage modules
+    </v-button>
 
     <p v-if="noModules">
       Modules let you keep track of the individual requirements and needs of your plant.
     </p>
 
-    <button type="button"
+    <v-button
       v-if="showTagButton"
-      class="footer-tags plain"
-      @click.prevent="emitShowTags">
-      <feather-hash />
-      <span>Add tags</span>
-    </button>
+      color="grey"
+      class="footer-tags"
+      @click.native.prevent="emitShowTags">
+      <feather-tag slot="icon" />
+      Add tags
+    </v-button>
   </section>
 </template>
 
 <script>
+  import Button from '@/components/Button'
   export default {
     name: 'PlantFooter',
 
@@ -32,10 +34,11 @@
     },
 
     components: {
+      'v-button': Button,
       'feather-box': () =>
         import('vue-feather-icons/icons/BoxIcon' /* webpackChunkName: "icons" */),
-      'feather-hash': () =>
-        import('vue-feather-icons/icons/HashIcon' /* webpackChunkName: "icons" */)
+      'feather-tag': () =>
+        import('vue-feather-icons/icons/TagIcon' /* webpackChunkName: "icons" */)
     },
 
     methods: {
@@ -56,7 +59,6 @@
     padding:
       calc(var(--base-gap) / 2) var(--base-gap)
       calc(var(--base-gap) / 2 + 3px);
-    font-size: var(--text-size-xsmall);
     font-style: italic;
     display: flex;
     justify-content: space-between;
@@ -80,6 +82,7 @@
 
     & p {
       margin: calc(var(--base-gap) * 1.5) 0;
+      font-size: var(--text-size-xsmall);
     }
   }
 </style>