+++ src/store/mutations/index.js
@@ -3,11 +3,13 @@ import notifications from './notifications'
 import plants from './plants'
 import tags from './tags'
 import settings from './settings'
+import version from './version'
 
 export default {
   ...appheader,
   ...notifications,
   ...plants,
   ...tags,
-  ...settings
+  ...settings,
+  ...version
 }