+++ CHANGELOG.md
@@ -5,7 +5,19 @@ All notable changes to this project will be documented here.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [1.2.0] - 2018-07-22
+
+### Added
+- Notification system to let users know a new version of the app has been released and downloaded.
+
+### Changed
+- Onboarding screen has been moved to its own route.
+- Renamed Changelog to Release Notes in Settings.
+- Updated icon for tags to be more clear.
+
+### Fixed
+- Some minor fixes with the button component and icons.
+- An issue with the build tools (PostCSS).
 
 ## [1.1.0] - 2018-07-16